+++ CHANGELOG.md
@@ -48,6 +48,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - UI
     - Date field full width
     - Add water meter reading modal layout
+    - Rent table overflow
 
 ## v0.2.0 - 2024-08-21